Load 3.3.30 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@113 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/main.c
@@ -311,7 +311,7 @@ main (int argc, char **argv)
         if (!erq_file)
         {
             fatal("Out of memory for erq pathname (%lu bytes).\n"
-                 , strlen(BINDIR)+6);
+                 , (unsigned long)strlen(BINDIR)+6);
         }
         strcpy(erq_file, BINDIR);
         strcat(erq_file, "/erq");
@@ -323,7 +323,7 @@ main (int argc, char **argv)
         if (!tmp)
         {
             fatal("Out of memory for erq pathname (%lu bytes).\n"
-                 , strlen(BINDIR)+2+strlen(erq_file));
+                 , (unsigned long)(strlen(BINDIR)+2+strlen(erq_file)));
         }
         strcpy(tmp, BINDIR);
         strcat(tmp, "/");
